Trait tokio_trace_core::field::ValidLen

pub trait ValidLen<'a>: Sealed + Borrow<[(&'a Field, Option<&'a (dyn Value + 'a)>)]> { }

Marker trait implemented by arrays which are of valid length to construct a ValueSet.

ValueSets may only be constructed from arrays containing 32 or fewer elements, to ensure the array is small enough to always be allocated on the stack. This trait is only implemented by arrays of an appropriate length, ensuring that the correct size arrays are used at compile-time.